[refactor]: 필터페이지 디테일페이지 마이페이지 컴포넌트 분리 및 코드 효율성 수정
필터페이지 디테일페이지 마이페이지 컴포넌트 분리 및 코드 효율성 수정
