+++ src/api/user.js
@@ -1,4 +1,3 @@
-import ourAxios from './ourAxios';
 import axios from 'axios';
 
 export const register = async ({username, password, email, phone}) => {
