+++ src/pages/FrameCustomPage.jsx
@@ -1,10 +1,9 @@
 import React from 'react';
 import FrameCustom from '../components/frameCustomPage/FrameCustom';
+
 const FrameCustomPage = () => {
     return (
-        <div>
-            <FrameCustom></FrameCustom>
-        </div>
+        <FrameCustom />
     );
 };
