+++ src/redux/modules/Filter.js
@@ -2,7 +2,6 @@ import { createSlice } from "@reduxjs/toolkit";
 
 // 초기값 설정
 const initialState = {
-    filterName:'test',
     blur:0,
     brightness:0,
     saturate:0,
@@ -18,6 +17,7 @@ const FilterSlice = createSlice({
   reducers: {
     SetFilter: (state, action) => {
       console.log(action.payload);
+      console.log(state);
       console.log({...state,...action.payload});
       return { ...state, ...action.payload }
     }
